backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.2
python-chess==1.999
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20
httpx==0.28.1
python-telegram-bot==21.10
